loadfiles class
