+++ Gotowy_system/STD_OS/src/stdos/Interface/SwitchInput.java
@@ -47,7 +47,7 @@ class SwitchInput {
                 case "taskcreate":
                     try {
                         ProcessManager.KM_CreateProcess(String.valueOf(arguments[1]), String.valueOf(arguments[2]), Integer.parseInt(arguments[3]));
-                        System.out.println("Utworzono plik: "+ arguments[1] +" proces: "+ arguments[2] + "priorytet: " +arguments[3]);
+                        System.out.println("Utworzono plik: "+ arguments[1] +" proces: "+ arguments[2] + " priorytet: " +arguments[3]);
                     }
                     catch(Exception e){
                         System.out.println("Blad Process Manager: polecenie {taskcreate}");
